refactor(Role.java): 공통된 로직이 아닌 메서드 삭제
특정 하위 클래스에서만 사용되는 메서드 삭제

BREAKING CHANGE: distributeBettingAmount, recordCompeteResult 삭제
as-is:
public abstract class Role {
    ...

    private BettingAmount bettingAmount;

    public void distributeBettingAmount(final Outcome outcome) {
        if (winWithoutBlackJack(outcome)) {
            bettingAmount.giveTwoTimes();
        }
    }

    private boolean winWithoutBlackJack(final Outcome outcome) {
        return outcome == Outcome.VICTORY && !hand.isBlackJack();
    }

    public void recordCompeteResult(final Outcome outcome, final Role player) {
        if (outcome == Outcome.VICTORY) {
            final int finalIncome = bettingAmount.getTotalValue() + player.getCurrentIncome();
            bettingAmount = new BettingAmount(finalIncome, bettingAmount.getInitialValue());
            player.loseAll();
        }
    }

    private int getCurrentIncome() {
        return bettingAmount.getTotalValue();
    }

    private void loseAll() {
        bettingAmount.loseAll();
    }

    ...
}

to-be:
public abstract class Role {
    ...
    protected BettingAmount bettingAmount;

    ...

    public abstract void distributeBettingAmount(final Outcome outcome);

    public abstract void distributeBettingAmount(final Outcome outcome, final Role role);

    protected int getCurrentIncome() {
        return bettingAmount.getTotalValue();
    }

    protected void loseAll() {
        bettingAmount.loseAll();
    }
}

Player:
as-is:
public class Player extends Role {
    ...
}

to-be:
public class Player extends Role {
    ...

    @Override
    public void distributeBettingAmount(Outcome outcome) {
        if (winWithoutBlackJack(outcome)) {
            bettingAmount.giveTwoTimes();
        }
    }

    private boolean winWithoutBlackJack(final Outcome outcome) {
        return outcome == Outcome.VICTORY && !hand.isBlackJack();
    }

    @Override
    public void distributeBettingAmount(Outcome outcome, Role role) {
        throw new IllegalArgumentException(METHOD_ERROR);
    }

    ...
}

Dealer:
as-is:
public class Dealer extends Role {
    ...
}

to-be:
public class Dealer extends Role {
    ...

    @Override
    public void distributeBettingAmount(Outcome outcome) {
        throw new IllegalArgumentException(METHOD_ERROR);
    }

    @Override
    public void distributeBettingAmount(Outcome outcome, Role role) {
        if (outcome == Outcome.VICTORY) {
            final int finalIncome = bettingAmount.getTotalValue() + role.getCurrentIncome();
            bettingAmount = new BettingAmount(finalIncome, bettingAmount.getInitialValue());
            role.loseAll();
        }
    }

    ...
}
